+++ server/src/routes/mod.rs
@@ -1,4 +1,5 @@
 mod add;
+mod ws;
 
-pub use add::print_client;
-pub use add::{add_handler as add_handler};
+pub use add::{print_client, add_handler};
+pub use ws::{ws_handler};
